# Conflicts:
#	Shelfalytics.UI/src/app/pages/main/main.component.html

+++ Shelfalytics.API/Shelfalytics.API/Models/EquipmentReadingModel.cs
@@ -8,7 +8,7 @@ namespace Shelfalytics.API.Models
     public class EquipmentReadingModel
     {
         public string IMEI { get; set; }
-        public int Temperature { get; set; }
+        public decimal Temperature { get; set; }
         public IEnumerable<EquipmentDistanceReadingModel> DistanceSensors { get; set; }
         public bool IsPoweredOn { get; set; }
     }
